model/requirements.txt
numpy==1.23.4
pandas==1.5.1
scikit-surprise==1.1.3
jupyter==1.0.0